requirements.txt
sc-compression
Pillow
pylzham
zstandard
pyliblzfse
texture2ddecoder